Separate Link file responsabilities
