+++ js/viz/constants/links.js
@@ -1,146 +1,17 @@
-import { getContainer } from "../draw/app.js";
-import { drawBezierLink } from "../draw/link.js";
+import {
+  ParentLink,
+  DaughterLink,
+  MCRecoParticleAssociation,
+  MCRecoClusterParticleAssociation,
+  MCRecoTrackParticleAssociation,
+  Clusters,
+  Tracks,
+  Particles,
+  Vertex,
+  ParticleIDLink,
+} from "../objects/Link.js";
 
-const colors = {
-  "parents": "#AA0000",
-  "daughters": "#00AA00",
-  "mcreco": "#0000AA",
-  "tracks": "#AAAA00",
-  "clusters": "#00AAAA",
-  "particles": "#AA00AA",
-  "mcclusters": "#D8F1A0",
-  "mctracks": "#fe5e41",
-  "vertex": "#593746",
-  "particle-id": "#00FF00",
-};
-
-export class Link {
-  constructor(from, to) {
-    this.from = from;
-    this.to = to;
-    this.color = "#A00";
-    this.xShift = 0;
-  }
-
-  draw() {
-    drawBezierLink(this);
-  }
-
-  isVisible() {
-    const boxFrom = this.from;
-    const boxTo = this.to;
-
-    const fromX = boxFrom.x + boxFrom.width / 2;
-    const fromY = boxFrom.y + boxFrom.height;
-    const toX = boxTo.x + boxTo.width / 2;
-    const toY = boxTo.y;
-
-    const boxX = Math.min(fromX, toX);
-    const boxWidth = Math.abs(fromX - toX);
-    const boxY = Math.min(fromY, toY);
-    const boxHeight = Math.abs(fromY - toY);
-
-    const viewport = getContainer();
-
-    return (
-      viewport.right > boxX &&
-      viewport.left < boxX + boxWidth &&
-      viewport.bottom > boxY &&
-      viewport.top < boxY + boxHeight
-    );
-  }
-}
-
-class ParentLink extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["parents"];
-    this.xShift = 3;
-  }
-
-  draw() {
-    drawBezierLink(this, true);
-  }
-}
-
-class DaughterLink extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["daughters"];
-    this.xShift = -3;
-  }
-}
-
-class MCRecoParticleAssociation extends Link {
-  constructor(from, to, weight) {
-    super(from, to);
-    this.color = colors["mcreco"];
-    this.weight = weight;
-  }
-
-  // draw(ctx) {
-  //   drawStraightLink(ctx, this);
-  // }
-}
-
-class Particles extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["particles"];
-  }
-}
-
-class Clusters extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["clusters"];
-  }
-}
-
-class Tracks extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["tracks"];
-  }
-}
-
-class Vertex extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["vertex"];
-  }
-}
-
-class ParticleIDLink extends Link {
-  constructor(from, to) {
-    super(from, to);
-    this.color = colors["particle-id"];
-  }
-}
-
-class MCRecoTrackParticleAssociation extends Link {
-  constructor(from, to, weight) {
-    super(from, to);
-    this.color = colors["mctracks"];
-    this.weight = weight;
-  }
-
-  // draw(ctx) {
-  //   drawStraightLink(ctx, this);
-  // }
-}
-
-class MCRecoClusterParticleAssociation extends Link {
-  constructor(from, to, weight) {
-    super(from, to);
-    this.color = colors["mcclusters"];
-    this.weight = weight;
-  }
-
-  // draw(ctx) {
-  //   drawStraightLink(ctx, this);
-  // }
-}
+export { Link } from "../objects/Link.js";
 
 export const linkTypes = {
   "parents": ParentLink,
