+++ js/viz/constants/styles.js
@@ -1,0 +1,12 @@
+export const linkColors = {
+  parents: "#AA0000",
+  daughters: "#00AA00",
+  mcreco: "#0000AA",
+  tracks: "#AAAA00",
+  clusters: "#00AAAA",
+  particles: "#AA00AA",
+  mcclusters: "#D8F1A0",
+  mctracks: "#fe5e41",
+  vertex: "#593746",
+  "particle-id": "#00FF00",
+};
